requirements.txt
groq==0.9.0
streamlit==1.35.0
pandas==2.2.2
numpy==1.26.4
scipy==1.13.0
vaderSentiment==3.3.2
plotly==5.22.0
python-dotenv==1.0.1
rich==13.7.1